Configure express-session
- Add new file : config/config.ts
  - has config for session()
  - has config for a MongoDBStore for sessions
  - moved constants from server.ts here
- Add notes on configuring express-session in : docs/notes.md
